Remove the git cleanup after publication.
Removing the cleanup is useful when working with cargo-sphinx in a Docker
container where your git credentials are not available. Not deleting the git
files means the push can be completed on the host machine instead.

The files are created beneath the _build output directory which is hosed on
the next cargo-sphinx call.